step_4

+++ app/src/main/java/hexlet/code/controller/UserController.java
@@ -3,7 +3,9 @@ package hexlet.code.controller;
 import hexlet.code.dto.UserCreateDTO;
 import hexlet.code.dto.UserDTO;
 import hexlet.code.dto.UserUpdateDTO;
+import hexlet.code.exception.AccessDeniedException;
 import hexlet.code.service.UserService;
+import hexlet.code.util.UserUtils;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -28,6 +30,9 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private UserUtils userUtils;
+
     @GetMapping(path = "")
     @ResponseStatus(HttpStatus.OK)
     public List<UserDTO> index() {
@@ -49,12 +54,18 @@ public class UserController {
     @PutMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.OK)
     public UserDTO update(@PathVariable Long id, @Valid @RequestBody UserUpdateDTO data) {
+        if (userUtils.getCurrentUser().getId() != id) {
+            throw new AccessDeniedException("You don't have permission to update this user");
+        }
         return userService.update(id, data);
     }
 
     @DeleteMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
+        if (userUtils.getCurrentUser().getId() != id) {
+            throw new AccessDeniedException("You don't have permission to delete this user");
+        }
         userService.delete(id);
     }
 }
